+++ lib/src/indexeddb_store.dart
@@ -18,147 +18,130 @@ part of lawndart;
  * Wraps the IndexedDB API and exposes it as a [Store].
  * IndexedDB is generally the preferred API if it is available.
  */
-class IndexedDbStore<V> extends Store<V> {
-  
+class IndexedDbStore extends Store {
+
   static Map<String, idb.Database> _databases = new Map<String, idb.Database>();
 
   final String dbName;
   final String storeName;
 
-  IndexedDbStore(this.dbName, this.storeName) : super._();
+  IndexedDbStore._(this.dbName, this.storeName) : super._();
+
+  static Future<IndexedDbStore> open(String dbName, String storeName) async {
+    var store = new IndexedDbStore._(dbName, storeName);
+    await store._open();
+    return store;
+  }
 
   /// Returns true if IndexedDB is supported on this platform.
   static bool get supported => idb.IdbFactory.supported;
 
-  Future open() {
+  Future _open() async {
     if (!supported) {
-      return new Future.error(
-        new UnsupportedError('IndexedDB is not supported on this platform'));
+      throw new UnsupportedError('IndexedDB is not supported on this platform');
     }
-    
+
     if (_db != null) {
       _db.close();
     }
-    
-    return window.indexedDB.open(dbName)
-    .then((idb.Database db) {
-      //print("Newly opened db $dbName has version ${db.version} and stores ${db.objectStoreNames}");
-      if (!db.objectStoreNames.contains(storeName)) {
-        db.close();
-        //print('Attempting upgrading $storeName from ${db.version}');
-        return window.indexedDB.open(dbName, version: db.version + 1,
-          onUpgradeNeeded: (e) {
-            //print('Upgrading db $dbName to ${db.version + 1}');
-            idb.Database d = e.target.result;
-            d.createObjectStore(storeName);
-          }
-        );
-      } else {
-        //print('The store $storeName exists in $dbName');
-        return db;
-      }
-    })
-    .then((db){
-      _databases[dbName] = db;
-      _isOpen = true;
-      return true;
-    });
+
+    var db = await window.indexedDB.open(dbName);
+
+    //print("Newly opened db $dbName has version ${db.version} and stores ${db.objectStoreNames}");
+    if (!db.objectStoreNames.contains(storeName)) {
+      db.close();
+      //print('Attempting upgrading $storeName from ${db.version}');
+      db = await window.indexedDB.open(dbName, version: db.version + 1,
+        onUpgradeNeeded: (e) {
+          //print('Upgrading db $dbName to ${db.version + 1}');
+          idb.Database d = e.target.result;
+          d.createObjectStore(storeName);
+        }
+      );
+    }
+
+    _databases[dbName] = db;
+    return true;
   }
-  
+
   idb.Database get _db => _databases[dbName];
 
   @override
-  Future _removeByKey(String  key) {
-    return _doCommand((idb.ObjectStore store) => store.delete(key));
+  Future removeByKey(String  key) {
+    return _runInTxn((store) => store.delete(key));
   }
 
   @override
-  Future _save(V obj, String key) {
-    return _doCommand((idb.ObjectStore store) {
-      return store.put(obj, key);
-    });
+  Future save(String obj, String key) {
+    return _runInTxn((store) => store.put(obj, key));
   }
 
   @override
-  Future<V> _getByKey(String key) {
-    return _doCommand((idb.ObjectStore store) => store.getObject(key),
-        'readonly');
+  Future<String> getByKey(String key) {
+    return _runInTxn((store) => store.getObject(key), 'readonly');
   }
 
   @override
-  Future _nuke() {
-    return _doCommand((idb.ObjectStore store) => store.clear());
+  Future nuke() {
+    return _runInTxn((store) => store.clear());
   }
 
-  Future _doCommand(Future requestCommand(idb.ObjectStore store),
-             [String txnMode = 'readwrite']) {
-    var completer = new Completer();
+  Future _runInTxn(Future requestCommand(idb.ObjectStore store),
+             [String txnMode = 'readwrite']) async {
     var trans = _db.transaction(storeName, txnMode);
     var store = trans.objectStore(storeName);
-    var future = requestCommand(store);
-    return trans.completed.then((_) => future);
+    var result = await requestCommand(store);
+    await trans.completed;
+    return result;
   }
 
-  Stream _doGetAll(dynamic onCursor(idb.CursorWithValue cursor)) {
-    var controller = new StreamController<V>();
+  Stream _doGetAll(dynamic onCursor(idb.CursorWithValue cursor)) async* {
     var trans = _db.transaction(storeName, 'readonly');
     var store = trans.objectStore(storeName);
-    // Get everything in the store.
-    store.openCursor(autoAdvance: true).listen(
-        (cursor) => controller.add(onCursor(cursor)),
-        onDone: () => controller.close(),
-        onError: (e) => controller.addError(e));
-    return controller.stream;
+    await for (var cursor in store.openCursor(autoAdvance: true)) {
+      yield onCursor(cursor);
+    }
   }
 
   @override
-  Stream<V> _all() {
+  Stream<String> all() {
     return _doGetAll((idb.CursorWithValue cursor) => cursor.value);
   }
 
   @override
-  Future _batch(Map<String, V> objs) {
-    var futures = <Future>[];
-
-    for (var key in objs.keys) {
-      var obj = objs[key];
-      futures.add(save(obj, key));
-    }
-
-    return Future.wait(futures);
-  }
-
-  @override
-  Stream<V> _getByKeys(Iterable<String> keys) {
-    var controller = new StreamController<V>();
-    Future.forEach(keys, (key) {
-      return getByKey(key).then((value) {
-        if (value != null) {
-          controller.add(value);
-        }
+  Future batch(Map<String, String> objs) {
+    return _runInTxn((store) {
+      objs.forEach((k, v) {
+        store.put(v, k);
       });
-    })
-    .then((_) => controller.close())
-    .catchError((e) => controller.addError(e));
-    return controller.stream;
-  }
-
-  @override
-  Future<bool> _removeByKeys(Iterable<String> keys) {
-    var completer = new Completer();
-    Future.wait(keys.map((key) => removeByKey(key))).then((_) {
-      completer.complete(true);
     });
-    return completer.future;
   }
 
   @override
-  Future<bool> _exists(String key) {
-    return getByKey(key).then((value) => value != null);
+  Stream<String> getByKeys(Iterable<String> keys) async* {
+    for (var key in keys) {
+      var v = await getByKey(key);
+      if (v != null) yield v;
+    }
   }
 
   @override
-  Stream<String> _keys() {
+  Future<bool> removeByKeys(Iterable<String> keys) {
+    return _runInTxn((store) {
+      for (var key in keys) {
+        store.delete(key);
+      }
+    });
+  }
+
+  @override
+  Future<bool> exists(String key) async {
+    var value = await getByKey(key);
+    return value != null;
+  }
+
+  @override
+  Stream<String> keys() {
     return _doGetAll((idb.CursorWithValue cursor) => cursor.key);
   }
 }